Change Balance type from Double to Int to keep it simple and focus on the intent of the lesson rather than the implementation.

+++ 01-basic/src/main/scala/dagmendez/basic/typeAliases.scala
@@ -4,4 +4,4 @@ object typeAliases:
 
   type Name    = String
   type IBAN    = String // International Bank Account Number
-  type Balance = Double
+  type Balance = Int
